Refactor of context code to make extension cleaner
